+++ R/expect-self-test.R
@@ -26,6 +26,36 @@ capture_success_failure <- function(expr) {
   )
 }
 
+format_success_failure <- function(status, exp_n_success, exp_n_failure) {
+  pluralise <- function(n, singular, plural) {
+    paste(n, ngettext(n, singular, plural))
+  }
+
+  tick <- cli::col_green(cli::symbol$tick)
+  cross <- cli::col_red(cli::symbol$cross)
+
+  success_ok <- status$n_success == exp_n_success
+  failure_ok <- status$n_failure == exp_n_failure
+
+  c(
+    sprintf(
+      "Expected %s and %s.",
+      pluralise(exp_n_success, "success", "successes"),
+      pluralise(exp_n_failure, "failure", "failures")
+    ),
+    sprintf(
+      "%s Observed %s.",
+      if (success_ok) tick else cross,
+      pluralise(status$n_success, "success", "successes")
+    ),
+    sprintf(
+      "%s Observed %s.",
+      if (failure_ok) tick else cross,
+      pluralise(status$n_failure, "failure", "failures")
+    )
+  )
+}
+
 #' Test your custom expectations
 #'
 #' @description
@@ -44,17 +74,11 @@ capture_success_failure <- function(expr) {
 expect_success <- function(expr) {
   status <- capture_success_failure(expr)
 
-  expected <- "Expected exactly one success and no failures."
-  if (status$n_success != 1) {
-    actual <- sprintf("Actually succeeded %i times", status$n_success)
-    fail(c(expected, actual))
-  } else if (status$n_failure > 0) {
-    actual <- sprintf("Actually failed %i times", status$n_failure)
-    fail(c(expected, actual))
-  } else {
+  if (status$n_success == 1 && status$n_failure == 0) {
     pass()
+    return(invisible())
   }
-
+  fail(format_success_failure(status, exp_n_success = 1, exp_n_failure = 0))
   invisible()
 }
 
@@ -63,21 +87,17 @@ expect_success <- function(expr) {
 expect_failure <- function(expr, message = NULL, ...) {
   status <- capture_success_failure(expr)
 
-  expected <- "Expected exactly one failure and no successes."
-  if (status$n_failure != 1) {
-    actual <- sprintf("Actually failed %i times", status$n_failure)
-    fail(c(expected, actual))
-  } else if (status$n_success != 0) {
-    actual <- sprintf("Actually succeeded %i times", status$n_success)
-    fail(c(expected, actual))
-  } else {
+  if (status$n_failure == 1 && status$n_success == 0) {
     if (is.null(message)) {
       pass()
     } else {
       act <- labelled_value(status$last_failure$message, "failure message")
       expect_match_(act, message, ..., title = "message")
     }
+    return(invisible())
   }
+
+  fail(format_success_failure(status, exp_n_success = 0, exp_n_failure = 1))
   invisible()
 }