+++ NEWS.md
@@ -1,6 +1,8 @@
 # testthat 3.3.1
 
 * `expect_snapshot()` now reports the original error class for base errors, rather than `rlang_error` (#2286).
+* `expect_success()` and `expect_failure()` are more clear about what the expectation actually did (#2297).
+* The hint to use `snapshot_download_gh()` is now only emitted when running in a job named "R-CMD-check" (#2300).
 * `expect_snapshot_file()` correctly reports file name if duplicated (@MichaelChirico, #2296).
 * Fixed support for `shinytest2::AppDriver$expect_values()` screenshot snapshot failing on CI (#2293, #2288).