+++ NEWS.md
@@ -3,6 +3,11 @@
 * `test_dir()` will no longer run tests in parallel if only a single file is being tested (#2305). 
 * `default_parallel_reporter()` is no longer exported; use `default_reporter(parallel = TRUE)` instead  (#2305).
 * New `LlmReporter()` designed for LLMs to read. It's currently used automatically inside Claude code, cursor, and gemini cli, and you can set `AGENT=1` to use with any coding agent (#2287).
+* `local_mocked_s3_method()` and `local_mocked_s4_method()` can now mock methods that don't already exist, and can use `definition = NULL` to temporarily remove a method. `local_mocked_s4_method()` now also works when the generic is defined in another package (#2302).
+* `expect_snapshot()` now reports the original error class for base errors, rather than `rlang_error` (#2286).
+* `expect_success()` and `expect_failure()` are more clear about what the expectation actually did (#2297).
+* The hint to use `snapshot_download_gh()` is now only emitted when running in a job named "R-CMD-check" (#2300).
+* `expect_snapshot_file()` correctly reports file name if duplicated (@MichaelChirico, #2296).
 * Fixed support for `shinytest2::AppDriver$expect_values()` screenshot snapshot failing on CI (#2293, #2288).
 
 * testthat now emits OpenTelemetry traces for tests when tracing is enabled. Requires the otel and otelsdk packages (#2282).